PAM, ATG, or compatible enzymes
CRISPR-Cas9 mechanisms recognize DNA targets that are complementary to a short CRISPR RNA (crRNA) sequence. The part of the crRNA sequence that is complementary to the target sequence is known as a spacer. In order for Cas9 to function, it also requires a specific protospacer adjacent motif (PAM) that varies depending on the bacterial species of the Cas9 gene.